Move Tornado code from OGFILE3.cpp into new file OGFILE_TORNADO.cpp.
