+++ include/OTORNADO.h
@@ -42,7 +42,6 @@
 
 //----------- Define class Tornado -----------//
 
-#pragma pack(1)
 class Tornado : public Sprite
 {
 public:
@@ -70,7 +69,6 @@ public:
 	int 	write_file(File* filePtr);
 	int	read_file(File* filePtr);
 };
-#pragma pack()
 
 //------- Define class TornadoArray ---------//
