+++ src/OGFILE3.cpp
@@ -32,10 +32,8 @@
 #include <OSITE.h>
 #include <OSNOWG.h>
 #include <OSPY.h>
-#include <OTORNADO.h>
 #include <OTOWN.h>
 #include <file_io_visitor.h>
-#include <visit_sprite.h>
 #include <dbglog.h>
 
 using namespace FileIOVisitor;
@@ -993,139 +991,6 @@ static void read_ai_info(File* filePtr, short** aiInfoArrayPtr, short& aiInfoCou
 	filePtr->file_read( *aiInfoArrayPtr, sizeof(short) * aiInfoCount );
 }
 //----------- End of static function read_ai_info ---------//
-
-//*****//
-
-//-------- Start of function TornadoArray::write_file -------------//
-//
-int TornadoArray::write_file(File* filePtr)
-{
-	filePtr->file_put_short(restart_recno);  // variable in SpriteArray
-
-	int    i;
-   Tornado *tornadoPtr;
-
-   filePtr->file_put_short( size() );  // no. of tornados in tornado_array
-
-   for( i=1; i<=size() ; i++ )
-   {
-      tornadoPtr = (Tornado*) get_ptr(i);
-
-      //----- write tornadoId or 0 if the tornado is deleted -----//
-
-      if( !tornadoPtr )    // the tornado is deleted
-      {
-         filePtr->file_put_short(0);
-      }
-      else
-      {
-         filePtr->file_put_short(1);      // there is a tornado in this record
-
-         //------ write data in the base class ------//
-
-         if( !tornadoPtr->write_file(filePtr) )
-            return 0;
-      }
-   }
-
-   //------- write empty room array --------//
-
-   write_empty_room(filePtr);
-
-   return 1;
-}
-//--------- End of function TornadoArray::write_file -------------//
-
-
-//-------- Start of function TornadoArray::read_file -------------//
-//
-int TornadoArray::read_file(File* filePtr)
-{
-	restart_recno    = filePtr->file_get_short();
-
-   int     i, tornadoRecno, tornadoCount;
-   Tornado* tornadoPtr;
-
-   tornadoCount = filePtr->file_get_short();  // get no. of tornados from file
-
-   for( i=1 ; i<=tornadoCount ; i++ )
-   {
-      if( filePtr->file_get_short() == 0 )
-      {
-         add_blank(1);     // it's a DynArrayB function
-      }
-      else
-      {
-         //----- create tornado object -----------//
-
-         tornadoRecno = tornado_array.create_tornado();
-         tornadoPtr   = tornado_array[tornadoRecno];
-
-         //----- read data in base class --------//
-
-         if( !tornadoPtr->read_file( filePtr ) )
-            return 0;
-      }
-   }
-
-   //-------- linkout() those record added by add_blank() ----------//
-   //-- So they will be marked deleted in DynArrayB and can be -----//
-   //-- undeleted and used when a new record is going to be added --//
-
-   for( i=size() ; i>0 ; i-- )
-   {
-      DynArrayB::go(i);             // since TornadoArray has its own go() which will call GroupArray::go()
-
-      if( get_ptr() == NULL )       // add_blank() record
-         linkout();
-   }
-
-   //------- read empty room array --------//
-
-   read_empty_room(filePtr);
-
-   return 1;
-}
-//--------- End of function TornadoArray::read_file ---------------//
-
-template <typename Visitor>
-static void visit_tornado(Visitor *v, Tornado *t)
-{
-	visit_sprite_members(v, t);
-   visit<float>(v, &t->attack_damage);
-   visit<int16_t>(v, &t->life_time);
-   visit<int16_t>(v, &t->dmg_offset_x);
-   visit<int16_t>(v, &t->dmg_offset_y);
-}
-
-enum { TORNADO_RECORD_SIZE = 44 };
-
-//--------- Begin of function Tornado::write_file ---------//
-//
-int Tornado::write_file(File* filePtr)
-{
-	return visit_with_record_size<FileWriterVisitor>(filePtr, this, &visit_tornado<FileWriterVisitor>,
-											TORNADO_RECORD_SIZE);
-}
-//----------- End of function Tornado::write_file ---------//
-
-//--------- Begin of function Tornado::read_file ---------//
-//
-int Tornado::read_file(File* filePtr)
-{
-	if (!visit_with_record_size<FileReaderVisitor>(filePtr, this, &visit_tornado<FileReaderVisitor>,
-										TORNADO_RECORD_SIZE))
-		return 0;
-
-   //------------ post-process the data read ----------//
-
-   sprite_info = sprite_res[sprite_id];
-
-	sprite_info->load_bitmap_res();
-
-	return 1;
-}
-//----------- End of function Tornado::read_file ---------//
 
 
 //*****//
